Add files via upload
I created this guessing game in the first week of my training with the basic concepts of python I learned. 
It's a program that lets the user choose from a list of 5 kinds of guessing games and get rewarded for guessing correctly. the program creates a csv file where the players data and activities are recorded.